Fuzzer refactoring: id generator interface and caching id generator
A custom id generator interface hierarchy is used instead of `IntSupplier`:
  * `IdGenerator` that allows to create fresh identifiers,
  * `IdentityPreservingIdGenerator`: `IdGenerator` that can return
    the same id for the same object.

A default implementation of `IdentityPreservingIdGenerator` is used in
fuzzer. It uses reference equality for object comparison, that allows
to create distinct models of equal object (in `equals` sense), and
to always assign the same id to the same enum value.